+++ test.js
@@ -3,22 +3,22 @@ const axios = require("axios");
 const useApi = require("./index");
 const { renderHook, act } = require("@testing-library/react-hooks");
 
-describe("performs GET requests", () => {
+describe("performs requests", () => {
   let mock;
   const url = "http://mock";
 
-  beforeAll(() => {
+  beforeEach(() => {
     mock = new MockAdapter(axios);
     jest.clearAllMocks();
     jest.useFakeTimers();
   });
 
-  afterAll(() => {
+  afterEach(() => {
     jest.useRealTimers();
     mock.restore();
   });
 
-  it("loads data from a url", async () => {
+  it("loads data from a url using GET", async () => {
     mock.onGet(url).reply(200, "response");
 
     const { result, waitForNextUpdate } = renderHook(() => useApi(url));
@@ -29,6 +29,147 @@ describe("performs GET requests", () => {
     await waitForNextUpdate();
 
     expect(result.current.data).toEqual("response");
+    expect(result.current.loading).toBeFalsy();
+  });
+
+  it("sends querystring data and loads data from a url using GET", async () => {
+    const params = { query: "hello" };
+    mock
+      .onGet(url)
+      .reply(config =>
+        config.params.query === "hello" ? [200, "response"] : [400, "error"]
+      );
+
+    const { result, waitForNextUpdate } = renderHook(() =>
+      useApi(url, 0, params)
+    );
+
+    expect(result.current.data).toEqual({});
+    expect(result.current.loading).toBeTruthy();
+
+    await waitForNextUpdate();
+
+    expect(result.current.data).toEqual("response");
+    expect(result.current.loading).toBeFalsy();
+  });
+
+  it("allows different formats of GET method param (GET)", async () => {
+    mock.onGet(url).reply(200, "response");
+
+    const { result, waitForNextUpdate } = renderHook(() =>
+      useApi(url, 0, { query: "hello" }, "GET")
+    );
+
+    await waitForNextUpdate();
+
+    expect(result.current.data).toEqual("response");
+    expect(result.current.loading).toBeFalsy();
+  });
+
+  it("allows different formats of GET method param (Get)", async () => {
+    mock.onGet(url).reply(200, "response");
+
+    const { result, waitForNextUpdate } = renderHook(() =>
+      useApi(url, 0, { query: "hello" }, "Get")
+    );
+
+    await waitForNextUpdate();
+
+    expect(result.current.data).toEqual("response");
+    expect(result.current.loading).toBeFalsy();
+  });
+
+  it("allows different formats of GET method param (gEt)", async () => {
+    mock.onGet(url).reply(200, "response");
+
+    const { result, waitForNextUpdate } = renderHook(() =>
+      useApi(url, 0, { query: "hello" }, "gEt")
+    );
+
+    await waitForNextUpdate();
+
+    expect(result.current.data).toEqual("response");
+    expect(result.current.loading).toBeFalsy();
+  });
+
+  it("loads data from a url using POST", async () => {
+    const postData = { query: "hello" };
+    mock.onPost(url, postData).reply(200, "response");
+
+    const { result, waitForNextUpdate } = renderHook(() =>
+      useApi(url, 0, postData, "post")
+    );
+
+    expect(result.current.data).toEqual({});
+    expect(result.current.loading).toBeTruthy();
+
+    await waitForNextUpdate();
+
+    expect(result.current.data).toEqual("response");
+    expect(result.current.loading).toBeFalsy();
+  });
+
+  it("allows different formats of POST method param (Post)", async () => {
+    const postData = { query: "hello" };
+    mock.onPost(url, postData).reply(200, "response");
+
+    const { result, waitForNextUpdate } = renderHook(() =>
+      useApi(url, 0, postData, "Post")
+    );
+
+    await waitForNextUpdate();
+
+    expect(result.current.data).toEqual("response");
+    expect(result.current.loading).toBeFalsy();
+  });
+
+  it("allows different formats of POST method param (POST)", async () => {
+    const postData = { query: "hello" };
+    mock.onPost(url, postData).reply(200, "response");
+
+    const { result, waitForNextUpdate } = renderHook(() =>
+      useApi(url, 0, postData, "POST")
+    );
+
+    await waitForNextUpdate();
+
+    expect(result.current.data).toEqual("response");
+    expect(result.current.loading).toBeFalsy();
+  });
+
+  it("allows different formats of POST method param (pOst)", async () => {
+    const postData = { query: "hello" };
+    mock.onPost(url, postData).reply(200, "response");
+
+    const { result, waitForNextUpdate } = renderHook(() =>
+      useApi(url, 0, postData, "pOst")
+    );
+
+    await waitForNextUpdate();
+
+    expect(result.current.data).toEqual("response");
+    expect(result.current.loading).toBeFalsy();
+  });
+
+  it("warns about bad finding a bad string in method type", async () => {
+    mock.onGet(url).reply(200, "response");
+
+    const { result, waitForNextUpdate } = renderHook(() =>
+      useApi(url, 0, { query: "hello" }, "POSTITNOIE")
+    );
+
+    expect(result.current.error).toBeTruthy();
+    expect(result.current.loading).toBeFalsy();
+  });
+
+  it("warns about garbage in method type", async () => {
+    mock.onGet(url).reply(200, "response");
+
+    const { result, waitForNextUpdate } = renderHook(() =>
+      useApi(url, 0, { query: "hello" }, { something: "wrong" })
+    );
+
+    expect(result.current.error).toBeTruthy();
     expect(result.current.loading).toBeFalsy();
   });