+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project are documented in this file.
 
+## Unreleased
+
+- Added: `payload` and `method` arguments.
+- Change: LICENSE copyright.
+
 ## 0.1.2
 
 - Added: npmignore